+++ CHANGES.md
@@ -25,6 +25,8 @@ Se agrega el módulo de administración y seguimiento de compras
     - Plazo en indicadores
     N [pages/modals/abmcompra.php]
         db: fecha_fin_contrato
+    - Alta proveedores
+    N [pages/helpers/abmroveedordb.php]
 
 
 ## FIX-VARIAS